try with Bun.serve + standalone Svelte
for now, we either have SSR'd pages but no reactivity (handmade plugin)
OR no SSR but reactivity (bun's official plugin)

thing is i would want to directly server-render from the DB, not have to
make a XHR roundtrip via API which would be dumb

another alterntive is a full sveltekit project, there's some interesting
work. main hurdle is sharing the drizzle schema, since we're using
bun:sqlite as the driver. Using better-sqlite3 and node adapter for
sveltekit results in loss of type info because the schema is not
type-compatible between DB drivers, but I saw a (unofficial) bun adapter
for sveltekit, so maybe we can use bun:sqlite on the svelte kit server
side anyways. im excited to try out sveltekit remote functions so maybe
ill go with that

but a standalone, Bun.served'd dashboard would be kinda neat (and
lightweight), and I might have the opportunity to contribute to
oven-sh/bun directly which would be nice, so idk. i feel like ssr for
the official plugin is achievalbe  (and its certainly not a roadmap
prority for the bun team)